+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/core/plugin/MQConsumerWrapper.java
@@ -75,4 +75,8 @@ public class MQConsumerWrapper extends MQWrapper {
     public void updateOffset(List<CloudEvent> events, AbstractContext eventMeshConsumeConcurrentlyContext) {
         meshMQPushConsumer.updateOffset(events, eventMeshConsumeConcurrentlyContext);
     }
+
+    public boolean isRetryEnabled() {
+        return meshMQPushConsumer.isRetryEnabled();
+    }
 }
