+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/core/protocol/http/consumer/EventMeshConsumer.java
@@ -150,15 +150,20 @@ public class EventMeshConsumer {
                     topic, event, subscriptionItem, eventMeshAsyncConsumeContext.getAbstractContext(),
                     consumerGroupConf, eventMeshHTTPServer, bizSeqNo, uniqueId, currentTopicConfig);
 
-                if (httpMessageHandler.handle(handleMsgContext)) {
-                    eventMeshAsyncConsumeContext.commit(EventMeshAction.ManualAck);
+                if (persistentMqConsumer.isRetryEnabled()) {
+                    boolean result = httpMessageHandler.syncHandle(handleMsgContext);
+                    eventMeshAsyncConsumeContext.commit(result ? EventMeshAction.CommitMessage : EventMeshAction.ReconsumeLater);
                 } else {
-                    try {
-                        sendMessageBack(event, uniqueId, bizSeqNo);
-                    } catch (Exception e) {
-                        //ignore
+                    if (httpMessageHandler.handle(handleMsgContext)) {
+                        eventMeshAsyncConsumeContext.commit(EventMeshAction.ManualAck);
+                    } else {
+                        try {
+                            sendMessageBack(event, uniqueId, bizSeqNo);
+                        } catch (Exception e) {
+                            //ignore
+                        }
+                        eventMeshAsyncConsumeContext.commit(EventMeshAction.CommitMessage);
                     }
-                    eventMeshAsyncConsumeContext.commit(EventMeshAction.CommitMessage);
                 }
             } finally {
                 TraceUtils.finishSpan(span, event);
