+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/core/protocol/http/push/HTTPMessageHandler.java
@@ -105,4 +105,9 @@ public class HTTPMessageHandler implements MessageHandler {
             return false;
         }
     }
+
+    @Override
+    public boolean syncHandle(HandleMsgContext handleMsgContext) {
+        return new SyncHTTPPushRequest(handleMsgContext).tryHTTPRequest();
+    }
 }
