+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/core/protocol/http/push/MessageHandler.java
@@ -25,4 +25,6 @@ import org.apache.eventmesh.runtime.core.protocol.http.consumer.HandleMsgContext
 public interface MessageHandler {
 
     boolean handle(HandleMsgContext handleMsgContext);
+
+    boolean syncHandle(HandleMsgContext handleMsgContext);
 }
