+++ eventmesh-starter/build.gradle
@@ -17,4 +17,6 @@
 
 dependencies {
     implementation project(":eventmesh-runtime")
+
+    implementation project(":eventmesh-storage-plugin:eventmesh-storage-rocketmq")
 }
