+++ eventmesh-storage-plugin/eventmesh-storage-api/src/main/java/org/apache/eventmesh/api/consumer/Consumer.java
@@ -43,4 +43,6 @@ public interface Consumer extends LifeCycle {
     void unsubscribe(String topic);
 
     void registerEventListener(EventListener listener);
+
+    boolean isRetryEnabled();
 }
