+++ eventmesh-storage-plugin/eventmesh-storage-kafka/src/main/java/org/apache/eventmesh/storage/kafka/consumer/KafkaConsumerImpl.java
@@ -87,6 +87,11 @@ public class KafkaConsumerImpl implements Consumer {
     }
 
     @Override
+    public boolean isRetryEnabled() {
+        return false;
+    }
+
+    @Override
     public synchronized void shutdown() {
         consumer.shutdown();
     }
