+++ eventmesh-storage-plugin/eventmesh-storage-knative/src/main/java/org/apache/eventmesh/storage/knative/consumer/KnativeConsumerImpl.java
@@ -70,6 +70,11 @@ public class KnativeConsumerImpl implements Consumer {
     }
 
     @Override
+    public boolean isRetryEnabled() {
+        return false;
+    }
+
+    @Override
     public void updateOffset(List<CloudEvent> cloudEvents, AbstractContext context) {
         pullConsumer.updateOffset(cloudEvents, context);
     }
