+++ eventmesh-storage-plugin/eventmesh-storage-mongodb/src/main/java/org/apache/eventmesh/storage/mongodb/consumer/MongodbConsumer.java
@@ -87,4 +87,9 @@ public class MongodbConsumer implements Consumer {
     public void registerEventListener(EventListener listener) {
         consumer.registerEventListener(listener);
     }
+
+    @Override
+    public boolean isRetryEnabled() {
+        return false;
+    }
 }
