+++ eventmesh-storage-plugin/eventmesh-storage-mongodb/src/main/java/org/apache/eventmesh/storage/mongodb/consumer/MongodbReplicaSetConsumer.java
@@ -119,6 +119,11 @@ public class MongodbReplicaSetConsumer implements Consumer {
         this.eventListener = listener;
     }
 
+    @Override
+    public boolean isRetryEnabled() {
+        return false;
+    }
+
     private void handle() {
         while (this.cursor.hasNext()) {
             ChangeStreamDocument<Document> next = cursor.next();
