+++ eventmesh-storage-plugin/eventmesh-storage-mongodb/src/main/java/org/apache/eventmesh/storage/mongodb/consumer/MongodbStandaloneConsumer.java
@@ -130,6 +130,11 @@ public class MongodbStandaloneConsumer implements Consumer {
         this.eventListener = listener;
     }
 
+    @Override
+    public boolean isRetryEnabled() {
+        return false;
+    }
+
     private FindIterable<Document> getCursor(MongoCollection<Document> collection, String topic, int lastId) {
         Document index = new Document("$gt", lastId);
         Document ts = new Document(MongodbConstants.CAPPED_COL_CURSOR_FN, index);
