+++ eventmesh-storage-plugin/eventmesh-storage-rabbitmq/src/main/java/org/apache/eventmesh/storage/rabbitmq/consumer/RabbitmqConsumer.java
@@ -133,6 +133,11 @@ public class RabbitmqConsumer implements Consumer {
         rabbitmqConsumerHandler.setEventListener(listener);
     }
 
+    @Override
+    public boolean isRetryEnabled() {
+        return false;
+    }
+
     public void setRabbitmqConnectionFactory(RabbitmqConnectionFactory rabbitmqConnectionFactory) {
         this.rabbitmqConnectionFactory = rabbitmqConnectionFactory;
     }
