+++ eventmesh-storage-plugin/eventmesh-storage-redis/src/main/java/org/apache/eventmesh/storage/redis/consumer/RedisConsumer.java
@@ -106,6 +106,11 @@ public class RedisConsumer implements Consumer {
         messageListener = new EventMeshMessageListener(listener);
     }
 
+    @Override
+    public boolean isRetryEnabled() {
+        return false;
+    }
+
     static class EventMeshMessageListener implements MessageListener<CloudEvent> {
 
         private final EventListener listener;
