+++ eventmesh-storage-plugin/eventmesh-storage-rocketmq/src/main/java/org/apache/eventmesh/storage/rocketmq/config/ClientConfiguration.java
@@ -95,4 +95,8 @@ public class ClientConfiguration {
     @ConfigFiled(field = "secretKey")
     @Builder.Default
     private String secretKey = "";
+
+    @ConfigFiled(field = "retryEnabled")
+    @Builder.Default
+    private boolean retryEnabled = false;
 }
