+++ eventmesh-storage-plugin/eventmesh-storage-rocketmq/src/main/java/org/apache/eventmesh/storage/rocketmq/consumer/RocketMQConsumerImpl.java
@@ -102,6 +102,11 @@ public class RocketMQConsumerImpl implements Consumer {
     }
 
     @Override
+    public boolean isRetryEnabled() {
+        return clientConfiguration.isRetryEnabled();
+    }
+
+    @Override
     public synchronized void shutdown() {
         pushConsumer.shutdown();
     }
