+++ eventmesh-storage-plugin/eventmesh-storage-rocketmq/src/main/resources/rocketmq-client.properties
@@ -19,3 +19,4 @@ eventMesh.server.rocketmq.namesrvAddr=127.0.0.1:9876;127.0.0.1:9876
 eventMesh.server.rocketmq.cluster=DefaultCluster
 eventMesh.server.rocketmq.accessKey=********
 eventMesh.server.rocketmq.secretKey=********
+eventMesh.server.rocketmq.retryEnabled=false
