+++ eventmesh-storage-plugin/eventmesh-storage-standalone/src/main/java/org/apache/eventmesh/storage/standalone/consumer/StandaloneConsumer.java
@@ -123,4 +123,9 @@ public class StandaloneConsumer implements Consumer {
     public void registerEventListener(EventListener listener) {
         this.listener = listener;
     }
+
+    @Override
+    public boolean isRetryEnabled() {
+        return false;
+    }
 }
