+++ eventmesh-storage-plugin/eventmesh-storage-standalone/src/main/java/org/apache/eventmesh/storage/standalone/consumer/StandaloneConsumerAdaptor.java
@@ -77,4 +77,9 @@ public class StandaloneConsumerAdaptor implements Consumer {
     public void registerEventListener(EventListener listener) {
         consumer.registerEventListener(listener);
     }
+
+    @Override
+    public boolean isRetryEnabled() {
+        return false;
+    }
 }
